+++ public/const.php
@@ -290,5 +290,6 @@ function fixUndefinedConstantsForCodeAnalysis()
         define('ENDPOINTS_SEARCH_PROTOCOL', '');
         define('ENDPOINTS_INDEXING_HOST', '');
         define('ENDPOINTS_INDEXING_TIMEOUT', 0);
+        define('DOI_EMAIL_CONTACT', '');
     }
 }